requirements.txt
Django==1.10.5
django-filter==1.0.1
djangorestframework==3.5.3
mysqlclient==1.3.10
tornado==4.4.2
social-auth-core==0.2.1
social-auth-app-django==0.1.0
jinja2==2.9.4
